parser: Fix memory leak in parsing sections
When we parse a token, we call strdup() on it in parser.l. In most
cases this is freed by mustache_free, but when parsing sections
we don't store the value of the closing tag name so the memory
reference gets lost. Because we don't use it anyway, it should
just be freed once we've finished building up that mustache_token_t.