feat: allow allocator customization at runtime
This adds a `mustache_memory_setup` function that can be used for custom
memory allocators like so:

```c
mustache_memory_setup(palloc, repalloc, palloc0, pfree);i
```

The main motivation is to be able to use mustache-c in postgresql
extensions. But this can be used in other ways too, like leak checking
with the Boehm GC.

This is completely optional and doesn't cause API breakage. mustache_c
keeps using the standard malloc, realloc and free by default.